分类删除的优化，分类结束。

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -11,13 +11,17 @@ import com.sky.entity.Category;
 import com.sky.entity.Employee;
 import com.sky.exception.AccountLockedException;
 import com.sky.exception.AccountNotFoundException;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.exception.PasswordErrorException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
 import com.sky.mapper.EmployeeMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
 import com.sky.service.EmployeeService;
 import lombok.extern.slf4j.Slf4j;
+import lombok.val;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -31,6 +35,10 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService {
     @Autowired
     CategoryMapper categoryMapper;
+    @Autowired
+    DishMapper dishMapper;
+    @Autowired
+    SetmealMapper setmealMapper;
 
     @Override
     public void save(CategoryDTO categoryDTO) {
@@ -69,7 +77,13 @@ public class CategoryServiceImpl implements CategoryService {
      */
     @Override
     public void updateCategory(CategoryDTO categoryDTO) {
-        categoryMapper.updateCategory(categoryDTO);
+        Category category = new Category();
+        BeanUtils.copyProperties(categoryDTO,category);
+        //更新人
+        category.setUpdateUser(BaseContext.getCurrentId());
+        //更新时间
+        category.setUpdateTime(LocalDateTime.now());
+        categoryMapper.updateCategory(category);
     }
 
     /**
@@ -88,6 +102,19 @@ public class CategoryServiceImpl implements CategoryService {
      */
     @Override
     public void deleteById(long id) {
+        ///查询当前分类是否关联了菜品，如果关联了就抛出业务异常
+        Integer count = dishMapper.countByCategoryId(id);
+        if(count > 0){
+            //当前分类下有菜品，不能删除，抛出一个删除异常
+            throw new  DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+
+        //查询当前分类是否关联了套餐，如果关联了就抛出业务异常
+        count = setmealMapper.countByCategoryId(id);
+        if(count > 0){
+            //当前分类下有套餐，不能删除，抛出一个删除异常
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
         categoryMapper.deleteById(id);
     }
 }
